Enhancement for clusterpool to work on on-prem clouds
Currently clusterpool is only available for public clouds like
aws, gcp and azure. Installing cluster on on-prem clouds require
additional pre-configurations that cannot be done on a fly as
currently implemented in clusterpool. Through this enhancemnent
we propose how these preconfigurations can be fed into the
clusterpool through a new `inventory` field.